+++ CHANGELOG.md
@@ -7,6 +7,12 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## Version 1.2.1 - Unreleased
 
+### Changed
+* Removed `16806_WORKAROUND` as it is not longer needed for Python 3.8+
+
+### Fixed
+* Fixed incorrect return type in docstrings
+
 
 ## Version 1.2.0 - Released 2024-08-20